+++ CHANGES.rst
@@ -12,6 +12,11 @@ In development
   and importing ``byte_bounds`` from ``np.lib.array_utils``.
   https://github.com/joblib/joblib/pull/1501
 
+- The parameter ``return_as`` in ``joblib.Parallel`` can now be set to
+  ``generator_unordered``. In this case the results will be returned in the
+  order of task completion rather than the order of submission.
+  https://github.com/joblib/joblib/pull/1463
+
 Release 1.3.2 -- 2023/08/08
 ---------------------------